[CHORE] Extract getRegularMaxDigs/getRemainingDigs into desert.ts
Move the two pure digs-left helpers out of the DesertDiggingDisplay HUD
component and into features/game/types/desert.ts, alongside the other
digging helpers (getTreasuresFound, getArtefactsFound, getTreasureCount).

No behaviour change — the functions are moved verbatim. This lets pure
consumers (e.g. tooling/overviews) compute remaining digs without pulling
the HUD component's React/GameProvider graph in. BeachScene and Digby now
import getRemainingDigs from desert.ts directly.
